[Refactor] : user 파일 안에 logIn과 signIn 구분 (#21)
