+++ src/main/java/com/potato/balbambalbam/user/signIn/controller/UserController.java
@@ -1,9 +1,9 @@
-package com.potato.balbambalbam.user.controller;
+package com.potato.balbambalbam.user.signIn.controller;
 
 import com.potato.balbambalbam.data.entity.User;
-import com.potato.balbambalbam.user.dto.UserDto;
+import com.potato.balbambalbam.user.signIn.dto.UserDto;
 import com.potato.balbambalbam.data.repository.UserRepository;
-import com.potato.balbambalbam.user.service.UserService;
+import com.potato.balbambalbam.user.signIn.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
